# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cakapku API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3456",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/perkins": {
            "get": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Get list of Perkin",
                "operationId": "47fe43230504b6aed6587cd29e71bc85",
                "responses": {
                    "200": {
                        "description": "List of Perkin"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Create new Perkin",
                "operationId": "158f1c33ea759242b4f59ce90bb7a471",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_perkin",
                                    "id_periode"
                                ],
                                "properties": {
                                    "nama_perkin": {
                                        "type": "string"
                                    },
                                    "no_sk": {
                                        "type": "string"
                                    },
                                    "id_periode": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Perkin created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perkins/{id}": {
            "put": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Update Perkin",
                "operationId": "d821a4629ccb0c3c6d20ed35b0aa8987",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_perkin": {
                                        "type": "string"
                                    },
                                    "no_sk": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perkin updated successfully"
                    },
                    "404": {
                        "description": "Perkin not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Delete Perkin",
                "operationId": "9c5eabff1813651368340aa69eba150e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perkin deleted successfully"
                    },
                    "404": {
                        "description": "Perkin not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perkins/{id}/assign-satker": {
            "post": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Assign Perkin to Satker",
                "operationId": "ef3c61f1e8b9b89a64a8c91a4f2a130d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_satkers"
                                ],
                                "properties": {
                                    "id_satkers": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assigned successfully"
                    },
                    "404": {
                        "description": "Perkin not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/perkins/import": {
            "post": {
                "tags": [
                    "Perkin"
                ],
                "summary": "Import Perkin and IKSK from Excel",
                "operationId": "b0413dabb5c2bcca9925b74f5b87c328",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id_periode": {
                                        "description": "ID of the associated Period",
                                        "type": "integer"
                                    },
                                    "file": {
                                        "description": "Excel File (.xlsx, .csv)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Import successful"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}